#42 SQLModel 적용
