+++ app/data_models/__init__.py
@@ -1,8 +1,1 @@
-from .member import Member
-from .plan import Plan
-from .spot_tag import SpotTag
-from .checklist import Checklist
-from .spot import Spot
-from .plan_spot_map import PlanSpotMap
-from .plan_spot_tag_map import PlanSpotTagMap
-from .administrative_division import AdministrativeDivision
+from .data_model import (AdministrativeDivision, Member, Plan, )
