+++ app/routers/oauths/naver_oauth_router.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter, Query, HTTPException, Response, Request
 from fastapi.responses import RedirectResponse
 
+from app.data_models.data_model import Member
+from app.repository.members.mebmer_repository import is_exist_member_by_email, save_member
 from app.services.oauths.naver_oauth_service import get_login_url, handle_callback, refresh_naver_access_token
 from app.utils.oauths.jwt_utils import create_jwt_naver
 
@@ -38,6 +40,16 @@ async def naver_callback(code: str, state: str, response: Response ):
             secure=False,
             samesite="None",
         )
+
+        # member정보 DB 저장
+        if not is_exist_member_by_email(user_info["email"]):
+            # 새 회원이면 DB저장
+            save_member(Member(
+                name=user_info["nickname"],
+                email=user_info["email"],
+                picture_url=user_info["profile_url"],
+                oauth="naver"
+            ))
 
         return {"content": "네이버 로그인 성공",
                     "nickname": user_info["nickname"],
